requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cssselect==1.2.0
environs==9.5.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
frozenlist==1.3.3
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
lxml==4.9.2
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
parse==1.19.0
pycparser==2.21
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
requests-html==0.10.0
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
